Fix harmless gcc 12 warning about adding enums
Add a cast to avoid -Wdeprecated-enum-enum-conversion from gcc 12.

No real changes.

(cherry picked from commit 82cb8aa859722d1bb62998e10185e477ebc33135)

+++ samples/notebook/notebook.cpp
@@ -329,7 +329,7 @@ MyFrame::MyFrame()
 #endif
     menuType->AppendRadioItem(ID_BOOK_SIMPLEBOOK, "&Simple book\tCtrl-7");
 
-    menuType->Check(ID_BOOK_NOTEBOOK + m_type, true);
+    menuType->Check(static_cast<int>(ID_BOOK_NOTEBOOK) + m_type, true);
 
     wxMenu *menuOrient = new wxMenu;
     menuOrient->AppendRadioItem(ID_ORIENT_DEFAULT, "&Default\tAlt-0");
